pyproject.toml
[tool.poetry]
name = "accounts_workshop"
version = "0.1.0"
description = ""
authors = ["Omar U. Espejel"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
starknet-py = "^0.18.2"
pyyaml = "^6.0.1"
termcolor = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
scripts = "scripts.main:deploy"
